getClientSessionCookie
